requirements.txt
Scrapy
awesome-slugify
